backend/rest_api/pyproject.toml
[tool.poetry]
name = "rest-api"
version = "0.1.0"
description = ""
authors = ["Yaroslav Pisarev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
aiofiles = "^24.1.0"
fastapi-login = "^1.10.3"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
qrcode = "^8.0"
requests = "^2.32.3"
pydantic-settings = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
